+++ lib/util.ts
@@ -1,7 +1,0 @@
-export interface Predicate<T> {
-  (value: T): boolean;
-}
-
-export function kComb<T>(value: T): () => T {
-  return () => value;
-}