+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## Added
+
+- More unit tests.
+
+## Changed
+
+- Clear separation of internal API.
+
 ## [0.1.1] - 2021-01-20
 
 ### Added